Update and rename luy_thua_nhi_phan.c to nhi_phan_luy_thua.c

+++ c/nhi_phan_luy_thua.c
@@ -2,16 +2,16 @@
 
 typedef long long ll;
 
-ll lt_np(ll, int);
+ll lt_np(ll, ll);
 
 int main() {
-    ll a; int b;
+    ll a; ll b;
     scanf("%lld %d", &a, &b);
     printf("%lld \n", lt_np(a, b));
     return 0;
 }
 
-ll lt_np(ll a, int b) {
+ll lt_np(ll a, ll b) {
     ll res = 1;
     while (b) {
         if (b%2) {
